refactor stacks add tests
