+++ stackformation/aws/stacks/base.py
